+++ loom/__init__.py
@@ -1,2 +1,2 @@
 """loom/__init__.py"""
-__version__ = "0.8.2"
+__version__ = "0.8.3"